Include functions to help fix cursor style
When using vim within a mintty shell on windows the cursor style is set
to a permanent vertical bar that is almost invisible an makes editing
very difficult.

This commit provides a couple of functions to help fix the cursor and
make it a blinking-block in normal mode and a blinking vertical bar in
edit mode.