Include information to disable certificate revocation check when downloading plug.vim
In some environments, plug.vim will fail to download due to ssl
certificate revocation checks.  These can be disabled, to allow the
download to succeed.  This commit provides an example command line that
can be used, rather than use this as the default since it reduces
security, and should only be used when confident.